With notes

+++ Lesson 2/H2P1.py
@@ -2,5 +2,5 @@ age = int(input("Enter your age:"))
 age_ten = age + 10
 print("In ten years you will be", age_ten, "years old.")
 ftemp = float(input("Enter the temperature in Fahrenheit:"))
-ctemp = ((ftemp - 32) * 5/9)
-print("The temperature in degrees Celsius is", ctemp )
+ctemp = float((ftemp - 32) * 5/9)
+print("The temperature in degrees Celsius is", ctemp)
